--- assets/components/manageContributionCta/manageContributionCta.scss
@@ -1,28 +1,0 @@
-.component-cta-link--manage-contribution {
-  background-color: gu-colour(garnett-neutral-1);
-  border: 1px solid gu-colour(garnett-neutral-1);
-  color: #fff;
-  height: 54px;
-  line-height: 54px;
-  padding: 0 30px 0 25px;
-  transition: background-color .3s;
-  margin-bottom: $gu-v-spacing * 3;
-  margin-top: $gu-v-spacing * 2;
-
-  @include mq($from: mobileMedium) {
-    width: 300px;
-  }
-}
-
-.component-manage-contribution-cta {
-  .svg-arrow-right-straight {
-    height: 35%;
-    top: 17px;
-    right: 13px;
-
-    // Fix for IE.
-    @include mq($from: mobileMedium) {
-      left: 321px;
-    }
-  }
-}
